+++ Magento/Mtf/Config/FileResolver/Mask.php
@@ -26,7 +26,7 @@ class Mask implements FileResolverInterface
      *
      * @param ModuleResolver|null $moduleResolver
      */
-    public function __construct($moduleResolver = null)
+    public function __construct(ModuleResolver $moduleResolver = null)
     {
         if ($moduleResolver) {
             $this->moduleResolver = $moduleResolver;